prompt_generation/pyproject.toml
[project]
name = "prompt-generation"
version = "1.2.0"
description = "AI character prompt generator for 2D→3D pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
    # Hunyuan 3D / Tencent Cloud dependencies
    "httpx>=0.27.0",
    "tencentcloud-sdk-python-ai3d>=3.0.0",
    "cos-python-sdk-v5>=1.9.0",
    # Dev/test dependencies
    "pytest>=8.0.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.21.0",
]

[project.scripts]
generate-prompts = "generate_prompts:app"
